+++ Symulator/ProblemInstance.cs
@@ -9,9 +9,9 @@ namespace Symulator
     public interface ProblemInstance
     {
         int[] GetInitialSolution();
-        int GetCost(int[] solution);
+        long GetCost(int[] solution);
         int[] Randomize(int[] solution); // zwaraca kopie oryginalnej tablicy - backup
         int[] ConvertToFinalSolution(int[] solution);
-        int compareOldNewSolution(int newS, int oldS); // 1 jeśli newS lepsze, -1 gorsze i 0 takie same
+        int compareOldNewSolution(long newS, long oldS); // 1 jeśli newS lepsze, -1 gorsze i 0 takie same
     }
 }
